Address PR review nits
- Rename "Per-Function Sampling" to "Per-Endpoint Sampling" in documentation
- Rename sampling_ratio parameter to sampling_ratio_override for clarity
- Move sampling logic to separate sampling.py module to isolate concerns
