+++ docs/docs/genai/tracing/prod-tracing.mdx
@@ -137,9 +137,9 @@ For high-volume applications, you may want to reduce the number of traces export
 
 The default value is `1.0`, which means all traces will be exported. When set to less than `1.0`, say `0.1`, only 10% of the traces will be exported. The sampling is done at the trace level, meaning that all spans in some traces will be exported or discarded together.
 
-#### Per-Function Sampling
+#### Per-Endpoint Sampling
 
-You can also configure sampling on a per-function basis using the `sampling_ratio` parameter on the `@mlflow.trace` decorator. This allows different functions to have different sampling rates, which is useful when you want to:
+You can also configure sampling on a per-endpoint basis using the `sampling_ratio_override` parameter on the `@mlflow.trace` decorator. This allows different endpoints to have different sampling rates, which is useful when you want to:
 
 - Sample high-volume, low-value endpoints at a lower rate (e.g., 10%)
 - Always trace critical transactions (100% sampling)
@@ -150,13 +150,13 @@ import mlflow
 
 
 # High-volume endpoint: sample only 10% of traces
-@mlflow.trace(sampling_ratio=0.1)
+@mlflow.trace(sampling_ratio_override=0.1)
 def high_volume_endpoint(request):
     return process_request(request)
 
 
 # Critical transaction: always trace
-@mlflow.trace(sampling_ratio=1.0)
+@mlflow.trace(sampling_ratio_override=1.0)
 def payment_processing(transaction):
     return process_payment(transaction)
 
@@ -169,10 +169,10 @@ def normal_endpoint(request):
 
 **Key behaviors:**
 
-- When `sampling_ratio` is specified on the decorator, it **overrides** the global `MLFLOW_TRACE_SAMPLING_RATIO` setting for that function
-- When `sampling_ratio` is not specified (or set to `None`), the global sampling ratio is used
-- `sampling_ratio` only applies at the trace level (root spans). Nested calls within an existing trace are always included, and their `sampling_ratio` settings are ignored.
-- The `sampling_ratio` must be a float between `0.0` and `1.0`
+- When `sampling_ratio_override` is specified on the decorator, it **overrides** the global `MLFLOW_TRACE_SAMPLING_RATIO` setting for that function
+- When `sampling_ratio_override` is not specified (or set to `None`), the global sampling ratio is used
+- `sampling_ratio_override` only applies at the trace level (root spans). Nested calls within an existing trace are always included, and their `sampling_ratio_override` settings are ignored.
+- The `sampling_ratio_override` must be a float between `0.0` and `1.0`
 
 ## Adding Context to Production Traces
